+++ Lilu/PrivateHeaders/kern_config.hpp
@@ -18,6 +18,7 @@ class Configuration {
 	 *  Possible boot arguments
 	 */
 	static constexpr const char *bootargOff {"-liluoff"};			// Disable the kext
+	static constexpr const char *bootargUserOff {"-liluuseroff"};	// Disable kext user patcher
 	static constexpr const char *bootargBeta {"-lilubeta"};			// Force enable the kext on unsupported os
 	static constexpr const char *bootargBetaAll {"-lilubetaall"};	// Force enable the kext and all plugins on unsupported os
 	static constexpr const char *bootargForce {"-liluforce"};		// Force enable the kext (including single user mode)
@@ -89,6 +90,11 @@ public:
 	 *  Disable the extension by default
 	 */
 	bool isDisabled {true};
+
+	/**
+	 *  User patcher is disabled on request
+	 */
+	bool isUserDisabled {false};
 	
 	/**
 	 *  Do not patch dyld shared cache unless asked