+++ Changelog.md
@@ -4,6 +4,7 @@ Lilu Changelog
 #### v1.2.6
 - Added Cannon Lake and Ice Lake definitions
 - Added missing typed getOSData APIs
+- Added `-liluuseroff` boot-arg to disable user patcher (for e.g. shared cache manipulation)
 - Switched to Apple lzvn implementation
 
 #### v1.2.5